requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
loguru==0.7.2
msgpack==1.1.0
multidict==6.1.0
nonebot-adapter-onebot==2.4.6
nonebot2==2.4.0
propcache==0.2.0
pydantic==2.10.1
pydantic_core==2.27.1
pygtrie==2.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
yarl==1.18.0
